page(settings): add new page settings with profile deletion
